pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mlx3d"
version = "0.0.1"
description = "Differentiable 3D rendering and geometry ops on Apple Silicon with MLX"
authors = ["AmirHossein Razlighi <arazlighi@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/amirhossein-razlighi/mlx3D"
repository = "https://github.com/amirhossein-razlighi/mlx3D"
keywords = ["3D", "graphics", "renderer", "differentiable", "MLX", "Apple Silicon", "Metal"]
packages = [{ include = "mlx3d", from = "src" }]


classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: MacOS :: MacOS X",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Multimedia :: Graphics"
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
mlx = { version = ">=0.10.0", markers = "platform_system == 'Darwin' and platform_machine == 'arm64'" }
numpy = ">=1.24"
pillow = ">=10.0"
typing-extensions = ">=4.8"


[tool.poetry.group.dev.dependencies]
ruff = ">=0.5"
mypy = ">=1.8"
pytest = ">=8.0"
pytest-cov = ">=4.1"

[tool.poetry.scripts]
mlx3d-hello = "mlx3d.renderer.hello:main"

[tool.ruff]
line-length = 100
target-version = "py311"
extend-select = ["I"]  # isort

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
